Fixed URL and Image issue

+++ app/Helpers/FileHelper.php
@@ -10,8 +10,8 @@ function uploadImage(string $entity, $oldFilePath, UploadedFile $file)
     $time    = time();
     $prefix  = '/uploads';
     $user_id = Auth::user()->id;
-
-    if(File::exists(public_path($oldFilePath))) {
+    
+    if(!empty($oldFilePath) && File::exists(public_path($oldFilePath))) {
         File::delete(public_path($oldFilePath));
     }
 
@@ -30,7 +30,7 @@ function uploadImage(string $entity, $oldFilePath, UploadedFile $file)
 
     $file_name = $time . '_' . $image_name;
     $image->move($destination_path, $file_name);
-    $imageURL = env('APP_URL') . $prefix . '/' . $path . '/' . $file_name;
+    $imageURL = $prefix . '/' . $path . '/' . $file_name;
     
     return $imageURL;
